Project protoype res files
The res files, based on android's preset Bottom Navigation Template. The important file containing the 5x5 setup is in /res/layout/fragment_dashboard.xml . Additionally you'll need to copy the /res/drawable/darktile.png  and /res/drawable/tile.png files as they were the drawables used to make the buttons. Naming convention for the grid of buttons is A-E for the rows (top to bottom) and 0-4 for the columns (left to right).